+++ src/Controller/GetStatusByIdsController.php
@@ -4,28 +4,32 @@ namespace App\Controller;
 
 use App\Enum\UserBookingStatusEnum;
 use App\Service\BookingServiceInterface;
+use App\Service\MetricsHelper;
 use App\Service\UserBookingCacheServiceInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpKernel\Attribute\AsController;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
-use Symfony\Component\Serializer\SerializerInterface;
 
 #[AsController]
 class GetStatusByIdsController extends AbstractController
 {
     public function __construct(
         private readonly BookingServiceInterface $bookingService,
-        private readonly SerializerInterface $serializer,
         private readonly UserBookingCacheServiceInterface $userBookingCacheService,
+        private readonly MetricsHelper $metricsHelper,
     ) {
     }
 
     public function __invoke(Request $request): JsonResponse
     {
+        $this->metricsHelper->incMethodTotal(__METHOD__, MetricsHelper::INVOKE);
+
         $exchangeIds = json_decode($request->getContent())->ids;
+
         if (empty($exchangeIds)) {
+            $this->metricsHelper->incExceptionTotal(NotFoundHttpException::class);
             throw new NotFoundHttpException('Resource not found');
         }
 
@@ -52,12 +56,16 @@ class GetStatusByIdsController extends AbstractController
                     $this->userBookingCacheService->changeCacheEntry($id, ['status' => $userBooking->status]);
                 }
             } catch (\Exception) {
+                $this->metricsHelper->incExceptionTotal(\Exception::class);
+
                 $statuses[] = [
                     'exchangeId' => $id,
                     'status' => null,
                 ];
             }
         }
+
+        $this->metricsHelper->incMethodTotal(__METHOD__, MetricsHelper::COMPLETE);
 
         return new JsonResponse($statuses);
     }