+++ src/EventListener/FailedMessageEventListener.php
@@ -7,6 +7,7 @@ use App\Message\CreateBookingMessage;
 use App\Message\SendBookingNotificationMessage;
 use App\Message\WebformSubmitMessage;
 use App\Repository\Resources\AAKResourceRepository;
+use App\Service\MetricsHelper;
 use App\Service\NotificationServiceInterface;
 use Symfony\Component\EventDispatcher\Attribute\AsEventListener;
 use Symfony\Component\Messenger\Event\WorkerMessageFailedEvent;
@@ -17,18 +18,26 @@ final class FailedMessageEventListener
     public function __construct(
         private readonly AAKResourceRepository $AAKResourceRepository,
         private readonly NotificationServiceInterface $notificationService,
+        private readonly MetricsHelper $metricsHelper,
     ) {
     }
 
     public function __invoke(WorkerMessageFailedEvent $event): void
     {
+        $this->metricsHelper->incMethodTotal(__METHOD__, MetricsHelper::INVOKE);
+
         $envelope = $event->getEnvelope();
         $message = $envelope->getMessage();
 
         if ($message instanceof WebformSubmitMessage) {
+            $webformId = $message->getWebformId();
+            $throwable = $event->getThrowable();
+
+            $message = sprintf('Failed to extract data (Error: %d) from webform with id: %s.', $throwable->getCode(), $webformId);
+
             $this->notificationService->notifyAdmin(
-                'Webform data retrieval failed.',
-                'Failed to extract data from webform.',
+                'Webform data retrieval failed',
+                $message,
                 null,
                 null
             );
@@ -55,5 +64,7 @@ final class FailedMessageEventListener
                 $resource
             );
         }
+
+        $this->metricsHelper->incMethodTotal(__METHOD__, MetricsHelper::COMPLETE);
     }
 }