+++ src/Service/NotificationService.php
@@ -39,12 +39,13 @@ class NotificationService implements NotificationServiceInterface
         private readonly MailerInterface $mailer,
         private readonly string $bindNotificationTimezone,
         private readonly string $bindNotificationDateFormat,
+        private readonly MetricsHelper $metricsHelper,
     ) {
         try {
             $this->validatedAdminNotificationEmail = $this->validationUtils->validateEmail(
                 $this->emailAdminNotification
             );
-        } catch (\Exception $exception) {
+        } catch (\Exception) {
             $this->logger->warning('No admin notification email set.');
         }
     }
@@ -54,6 +55,8 @@ class NotificationService implements NotificationServiceInterface
      */
     public function sendBookingNotification(Booking $booking, ?AAKResource $resource, NotificationTypeEnum $type): void
     {
+        $this->metricsHelper->incMethodTotal(__METHOD__, MetricsHelper::INVOKE);
+
         $data = [
             'booking' => $booking,
             'resource' => $resource,
@@ -67,6 +70,8 @@ class NotificationService implements NotificationServiceInterface
         $notification = $this->buildNotification($type, $data);
 
         $this->sendNotification($notification);
+
+        $this->metricsHelper->incMethodTotal(__METHOD__, MetricsHelper::COMPLETE);
     }
 
     /**
@@ -77,6 +82,8 @@ class NotificationService implements NotificationServiceInterface
         ?AAKResource $resource,
         NotificationTypeEnum $type
     ): void {
+        $this->metricsHelper->incMethodTotal(__METHOD__, MetricsHelper::INVOKE);
+
         $body = $userBooking->body;
 
         $crawler = new Crawler($body);
@@ -157,6 +164,8 @@ class NotificationService implements NotificationServiceInterface
         $notificationData['adminNotification'] = true;
 
         $this->sendNotification($notificationData);
+
+        $this->metricsHelper->incMethodTotal(__METHOD__, MetricsHelper::COMPLETE);
     }
 
     /**
@@ -209,6 +218,8 @@ class NotificationService implements NotificationServiceInterface
      */
     public function notifyAdmin(string $subject, string $message, ?Booking $booking, ?AAKResource $resource): void
     {
+        $this->metricsHelper->incMethodTotal(__METHOD__, MetricsHelper::INVOKE);
+
         if ($this->validatedAdminNotificationEmail) {
             $to = $this->validatedAdminNotificationEmail;
             $template = 'email-notify-admin.html.twig';
@@ -243,6 +254,8 @@ class NotificationService implements NotificationServiceInterface
 
             $this->sendNotification($notificationData);
         }
+
+        $this->metricsHelper->incMethodTotal(__METHOD__, MetricsHelper::COMPLETE);
     }
 
     /**