+++ tests/Handler/AddBookingToCacheHandlerTest.php
@@ -9,9 +9,9 @@ use App\MessageHandler\AddBookingToCacheHandler;
 use App\Repository\Resources\AAKResourceRepository;
 use App\Security\Voter\BookingVoter;
 use App\Service\BookingServiceInterface;
+use App\Service\MetricsHelper;
 use App\Service\UserBookingCacheServiceInterface;
 use App\Tests\AbstractBaseApiTestCase;
-use Psr\Log\LoggerInterface;
 
 class AddBookingToCacheHandlerTest extends AbstractBaseApiTestCase
 {
@@ -27,8 +27,6 @@ class AddBookingToCacheHandlerTest extends AbstractBaseApiTestCase
             ->getMock();
         $userBookingCacheServiceMock->expects($this->exactly(1))->method('addCacheEntryFromArray');
 
-        $loggerMock = $this->createMock(LoggerInterface::class);
-
         $res = new AAKResource();
         $res->setResourceDisplayName('Cool resource name');
 
@@ -38,11 +36,13 @@ class AddBookingToCacheHandlerTest extends AbstractBaseApiTestCase
             ->getMock();
         $aakResourceRepositoryMock->expects($this->exactly(1))->method('findOneBy')->willReturn($res);
 
+        $metric = $this->createMock(MetricsHelper::class);
+
         $handler = new AddBookingToCacheHandler(
             $bookingServiceMock,
             $userBookingCacheServiceMock,
-            $loggerMock,
             $aakResourceRepositoryMock,
+            $metric,
         );
 
         $booking = new Booking();